Add dotfiles support for debian based systems

+++ brew/init.zsh
@@ -1,6 +1,6 @@
 #!/bin/bash
 BREW_FILE="${HOME}/dotfiles/brew/Brewfile"
-if [ -e ${BREW_FILE} ] 
+if [ -e ${BREW_FILE} ] && [ `uname` == 'Darwin' ]
 then
   brew bundle install --file=${BREW_FILE}
 fi
